requirements.txt
pyrogram
tgcrypto
python-dotenv
rich
prompt_toolkit
